requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cloudinary==1.26.0
config==0.5.1
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==3.2.8
django-bootstrap3==15.0.0
django-heroku==0.3.1
django-registration==3.2
django-registration-redux==2.9
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==8.4.0
psycopg2==2.9.1
python-decouple==3.5
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
